requirements.txt
Flask==2.0.1
Pillow==8.4.0
numpy==1.21.0
opencv-python==4.11.0.86
gunicorn==20.1.0  # Solo per il deployment in produzione 